pyproject.toml
[tool.poetry]
name = "knewnew_project"
version = "0.1.0"
description = ""
authors = ["Ahnsang0915 <tmqjf4921@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.13"
Django = "^4.1"
djangorestframework = "^3.13.1"
python-dotenv = "^0.20.0"
psycopg2 = "^2.9.3"
djangorestframework-simplejwt = "^5.2.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pylint = "^2.14.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
